Tweaks and more tests
