+++ src/NuGet.Services.Entities/FederatedCredentialPolicy.cs
@@ -14,6 +14,8 @@ namespace NuGet.Services.Entities
     /// </summary>
     public class FederatedCredentialPolicy : IEntity
     {
+        public const int MaxPolicyNameLength = 64;
+
         /// <summary>
         /// The unique key for this federated credential policy. Generated by the database.
         /// </summary>
@@ -58,7 +60,7 @@ namespace NuGet.Services.Entities
         /// <summary>
         /// User provided policy name to help identify this policy. This is not used in any evaluation of the policy.
         /// </summary>
-        [StringLength(64)]
+        [StringLength(MaxPolicyNameLength)]
         public string PolicyName { get; set; }
 
         /// <summary>
