+++ src/NuGetGallery/ViewModels/TrustedPublisherPolicyViewModel.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.ComponentModel.DataAnnotations;
 using System.Diagnostics;
 
 namespace NuGetGallery
@@ -11,17 +12,30 @@ namespace NuGetGallery
     [DebuggerDisplay("{PolicyName,nq}")]
     public sealed class TrustedPublisherPolicyViewModel
     {
+        private string _policyName = string.Empty;
+        private string _owner = string.Empty;
+
         public int Key { get; set; }
 
         /// <summary>
         /// User provided policy name.
         /// </summary>
-        public string PolicyName { get; set; }
+        [Required]
+        public string PolicyName
+        {
+            get => _policyName;
+            set => _policyName = value?.Trim() ?? string.Empty;
+        }
 
         /// <summary>
         /// NuGet package owner.
         /// </summary>
-        public string Owner { get; set; }
+        [Required]
+        public string Owner
+        {
+            get => _owner;
+            set => _owner = value?.Trim() ?? string.Empty;
+        }
 
         public TrustedPublisherPolicyInvalidReason? InvalidReason { get; set; }
